handle the control word and mode selection correctly.
this flips the bit that makes the pvt mode actually go as well as
setting the right mode when you add a pvt segment. I've also included
a startPvtMove and a stopPvtMove method. the intended use is to
add at least two pvt segments, then call startPvtMove. to stop
you can either add segments that stop, call stopPvtMove and specify
a duration to reach zero velocity, or call stopPvtMove and set the
optional emergency bool to true which immediately flips the control
word bit that stops the servo.