added position and time code to the WheelPod class.
this lets you specify a series of segments that define steering
angle, steering speed, wheel distance and wheel speed in actual units.
the wheel pod code then converts those units in to servo-friendly
units and passes them on to the PVT code I recently added.

note that there's a hard limit on how long segments for the servo
can be: 255 milliseconds. If you pass a segment in that exceeds
that duration, the wheel pod code will attempt to break it up into
smaller segments, but the way it does so may cause problems.
it is probably better to assume responsibility for the segment
length and keep the segments short enough to not hit this bit of code.