watch the buffer depth on the servos.
when we've added enough data to push the buffer beyond lenth 2,
tell the servos to start moving. it is unclear if you need to do this
after an underflow, but you definitely need to do this the first
time you start moving. hopefully doing this automatically isn't a
terrible plan.