scary test mode!
untested. it does compile. more or less scary? you decide!!
the point of this mode is to allow us to test the PVT control and
make sure all the actions are synchronized correctly. for sure don't
enable this mode unless the robot is all the way up on blocks!
like really, that would be very bad.