debug print statements and commented out calls to send pvt segment
all the recovering from error code was commented out and that made
things actually work. I should either consider removing this code
all the way or figuring out why it is broken. it was written with
assumptions about the way we were going to use the buffer that are
no longer true.