be a lot more careful about exactly how we start moving.
now keep track of servo buffer readiness (doesn't actually work,
but it is a nice idea) and then make Motion figure out when to
actually start. making Motion call for the start is a good idea,
but because the servos don't report buffer status until after they're
told to start moving, the readiness is a bit of a bust. it should
be refactored into something more useful or just dropped all together.