hacks and terminal output.
this works around the annoying "no buffer state 'till we're already
moving" problem by just waiting some number of sync messages and going
for it. this is probably not an ideal way of making the robot go, so
something should be done.

also, there's enough error print statements in here to be really
annoying. they should probably go away...