getNeedsToStart now really works. also quieted a bunch of output
I now keep track of the status of the pvt mode enabled flag on the
servo. this lets me track how many things should have been sent to
the servo so I can then know when we've sent enough things to start
going and we aren't going currently. this is exactly the needs to start
flag! now you can reliably know when you should tell the servo to
actually start moving in pvt mode!