fixed bug in pvt mode status.
the bit flag I was using to determine if pvt mode was active is
actually a flag that gets reused for other things in other modes,
so if you don't check what mode you're in and read the flag, you're
very sad about the results. this should fix the race condition brandon
attempted to fix earlier.