added special case handling code for the turn in place stuff.
also changed tabs to spaces. merging this will be not fun.
the turn in place code is pretty much directly copied from russel's
spin.py script, including variable names and style, which makes for
a bit of an akward bit of C++. hopefully this works because debugging
won't be fun!!