now the driving test launch uses the smooth planner!
WATCH OUT! the smooth planner currently doesn't really have a costmap
to look at, so it will probably crash in to things if you're not
careful! I turned the speed way down so hopefully you'll have time
to react to this, but until contradict gets the costmap piped into
the right spot, be careful!