Make filterSpeckles faster
I do not understand why this is faster.
Switching from registerPageLocked -> CudaMem has no effect.
Using a simple C-style function pointer is the same as using all the C++
lambda magic with member functions.
The thing that actually made the speedup is waitForCompletion instead of
enqueueHostCallback. There are other HostCallbacks that do not seem to slow
things down, so I am stumped.