Add wstool dependency on image_transport_plugins
These depend on having the same OpenCV version as the other nodes used around
them, so we need to rebuild them with OpenCV3. I am trying a new way of adding a
dependency this time. wstool is a tool for building and maintaining ros
workspaces. The .rosinstall file added in this commit describes a set of SCM
checkouts to add to a workspace. Run
wstool update -t src
from the ros_workspace directory to add the new checkout. It should only be
necessary to do this when a new dependency is added.